Создан класс Employee, созданы контроллёр-сервис-репозиторий-страницы для этой сущности

+++ src/main/java/ru/gb/timesheet/controller/ProjectController.java
@@ -41,7 +41,7 @@ public class ProjectController {
     }
 
     @DeleteMapping("/{id}")
-    public  ResponseEntity<Void> delete(@PathVariable int id){
+    public  ResponseEntity<Void> delete(@PathVariable Long id){
         service.delete(id);
         return ResponseEntity.noContent().build();
     }
